Refactor Movie API
