+++ src/main/java/tr/org/ab/spring/rest/videostore/movie/service/MovieServiceImpl.java
@@ -1,13 +1,9 @@
-package tr.org.ab.spring.rest.videostore.movie;
+package tr.org.ab.spring.rest.videostore.movie.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.type.CollectionType;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import tr.org.ab.spring.rest.videostore.movie.Movie;
 
-import java.io.IOException;
 import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
@@ -15,24 +11,10 @@ import java.util.concurrent.ConcurrentHashMap;
 /**
  * @author Omer Ozkan
  */
-@Component
+@Service
 class MovieServiceImpl implements MovieService {
 
     private Map<String, Movie> movies = new ConcurrentHashMap<>();
-
-    public MovieServiceImpl() {
-        ObjectMapper mapper = new ObjectMapper();
-        List<Movie> list = null;
-        try {
-            CollectionType reference = mapper.getTypeFactory().constructCollectionType(List.class, Movie.class);
-            list = mapper.readValue(new ClassPathResource("movies.json").getFile(), reference);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        for (Movie movie : list) {
-            movies.put(movie.getId(), movie);
-        }
-    }
 
     @Override
     public Movie getMovie(String uuid) {
